+++ migrations/000005_create_project_model_allowlist.down.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+DROP TABLE app.project_logical_models;
+
+COMMIT;